VCV-235: separate active metric snapshot schema
