+++ src/api/user/get-all-user-active-metrics.ts
@@ -1,9 +1,9 @@
 import {
     getUserActiveMetricsQueryParamsSchema,
     getUserActiveMetricsResponseSchema,
-    type ActiveMetricSnapshot,
     type GetUserActiveMetricsResponseBody,
 } from '@/api/user/validation/get-user-active-metrics-schema';
+import type { ActiveMetricSnapshot } from '@/api/user/validation/active-metric-snapshot-schema';
 import {
     type GetAllUserActiveMetricsQueryParams,
     type GetAllUserActiveMetricsResponseBody,
