+++ src/api/user/validation/get-user-active-metrics-schema.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod';
 import { booleanQueryParamSchema } from '@/lib/network/validation/boolean-query-param-schema';
+import { activeMetricSnapshotSchema } from '@/api/user/validation/active-metric-snapshot-schema';
 
 export const getUserActiveMetricsQueryParamsSchema = z.object({
     startDate: z.string().optional(),
@@ -8,17 +9,6 @@ export const getUserActiveMetricsQueryParamsSchema = z.object({
     globalOnly: booleanQueryParamSchema().optional(),
     limit: z.coerce.number().min(1).max(365).optional(),
     offset: z.coerce.number().min(0).optional(),
-});
-
-export const activeMetricSnapshotSchema = z.object({
-    id: z.number(),
-    snapshotDate: z.string(),
-    programId: z.number().nullable(),
-    a1Count: z.number(),
-    a7Count: z.number(),
-    a30Count: z.number(),
-    createdAt: z.string(),
-    updatedAt: z.string(),
 });
 
 export const getUserActiveMetricsResponseSchema = z.object({
@@ -33,7 +23,6 @@ export const getUserActiveMetricsResponseSchema = z.object({
 export type GetUserActiveMetricsQueryParams = z.infer<
     typeof getUserActiveMetricsQueryParamsSchema
 >;
-export type ActiveMetricSnapshot = z.infer<typeof activeMetricSnapshotSchema>;
 export type GetUserActiveMetricsResponseBody = z.infer<
     typeof getUserActiveMetricsResponseSchema
 >;
